Automated alembic downgrade upgrade on all databases

+++ lib/rucio/db/sqla/migrate_repo/versions/1a80adff031a_create_index_on_rules_hist_recent.py
@@ -1,38 +1,46 @@
-'''
- Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+# - Martin Barisits <martin.barisits@cern.ch>, 2015
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
- Licensed under the Apache License, Version 2.0 (the "License");
- You may not use this file except in compliance with the License.
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+''' create index on rules_hist_recent '''
 
- Authors:
- - Martin Barisits, <martin.barisits@cern.ch>, 2015
- - Vincent Garonne, <vincent.garonne@cern.ch>, 2017
-
-Create index on rules_hist_recent
-
-Revision ID: 1a80adff031a
-Revises: ae2a56fcc89
-Create Date: 2015-03-20 14:52:53.013432
-'''
-
+from alembic import context
 from alembic.op import create_index, drop_index
 
 
-# revision identifiers, used by Alembic.
-revision = '1a80adff031a'  # pylint: disable=invalid-name
-down_revision = '3ad36e2268b0'  # pylint: disable=invalid-name
+# Alembic revision identifiers
+revision = '1a80adff031a'
+down_revision = '3ad36e2268b0'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    create_index('RULES_HIST_RECENT_SC_NA_IDX', 'rules_hist_recent', ['scope', 'name'])
+
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        create_index('RULES_HIST_RECENT_SC_NA_IDX', 'rules_hist_recent', ['scope', 'name'])
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_index('RULES_HIST_RECENT_SC_NA_IDX', 'rules_hist_recent')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        drop_index('RULES_HIST_RECENT_SC_NA_IDX', 'rules_hist_recent')
